+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - 2017-10-XX
+## [0.3.0] - 2017-10-22
 ### Added
-- Settings & account panels
+- Language, Country and Show Comments from All Languages settings
+- Replying to comments
 
 ### Changed
 - Fix for sorting hide/show functionality